avfilter/vf_pullup: use ptrdiff_t as stride argument for dsp functions
This should avoid issues on x86_64

(cherry picked from commit 6dffc8f5aaab6b20385f0a0d9ef95cec7d6cdd4b)

+++ libavfilter/x86/vf_pullup_init.c
@@ -23,9 +23,9 @@
 #include "libavutil/x86/cpu.h"
 #include "libavfilter/vf_pullup.h"
 
-int ff_pullup_filter_diff_mmx(const uint8_t *a, const uint8_t *b, int s);
-int ff_pullup_filter_comb_mmx(const uint8_t *a, const uint8_t *b, int s);
-int ff_pullup_filter_var_mmx (const uint8_t *a, const uint8_t *b, int s);
+int ff_pullup_filter_diff_mmx(const uint8_t *a, const uint8_t *b, ptrdiff_t s);
+int ff_pullup_filter_comb_mmx(const uint8_t *a, const uint8_t *b, ptrdiff_t s);
+int ff_pullup_filter_var_mmx (const uint8_t *a, const uint8_t *b, ptrdiff_t s);
 
 av_cold void ff_pullup_init_x86(PullupContext *s)
 {
